+++ app/ids-loading-indicator/index.js
@@ -1,0 +1,1 @@
+import IdsLoadingIndicator from '../../src/ids-loading-indicator/ids-loading-indicator';